Restricting wp-includes/*.php to internal requests only breaks tinymce.  Removing this block
TinyMCE breaks with WordPress doing some concatenation etc when wp-includes is restricted to internal requests only.  This block seems to always break something, so I'm just removing it until a better idea comes up.  In the end, restricting direct requests to wp-includes does not provide a ton of extra security, so I'm comfortable removing it.