Update wordpress_security.conf with redirect rather than deny
Instead of returning a 403 when a sensitive file is requested, redirect this traffic back to the homepage instead.  The reasoning is that by testing certain files and getting a 403 error, an attacker can see what files we determine to be sensitive, which can reveal data about what kind of CMS or server technology we are running.  Doing a redirect back to the homepage makes less of a statement about whether the information is sensitive or not, thereby exposing less information to the bad guys.